requirements.txt
azure-cognitiveservices-vision-computervision<=0.9.1
pillow<=10.4.0
openai<=1.61.1
inference_sdk<=0.41.0
python-dotenv<=1.0.1
msrest<=0.7.1
pillow-heif

